Actualiza

+++ test/dummy/db/structure.sql
@@ -1132,8 +1132,8 @@ CREATE TABLE public.sipd_dominio_operaen_pais (
 --
 
 CREATE TABLE public.sipd_dominio_usuario (
-    dominio_id bigint NOT NULL,
-    usuario_id bigint NOT NULL
+    dominio_id integer NOT NULL,
+    usuario_id integer NOT NULL
 );
